fix: changed plugin and parameter idexing to 1 based on new Ardour 5.5

+++ app/src/main/java/de/paraair/ardmix/ArdourPlugin.java
@@ -65,7 +65,7 @@ public  class ArdourPlugin {
         int flags;
         float min;
         float max;
-        float current;
+        double current;
         String print_fmt;
         int scaleSize;
